fixed 500 error when `hd` field missing on JWT
When accessing through the API Gateway, the JWT payload contains
different fields than when accessing directly through the IAP.

Eg, a normal direct access has a payload like:

```
{
  "aud": "/projects/466589242038/apps/anders-klanten-dev",
  "azp": "/projects/466589242038/apps/anders-klanten-dev",
  "email": "anders@noderabbit.com",
  "exp": 1638350499,
  "hd": "noderabbit.com",
  "iat": 1638349899,
  "iss": "https://cloud.google.com/iap",
  "sub": "accounts.google.com:108121296498532996032"
}
```

And we use the `email` field to determine which user it is and the
`hd` field to potentially grant them admin/superuser access.

When accessing via API Gateway, the gateway's service account
apparently does things a bit different and the payload looks more
like:

```
{
  "aud": "/projects/466589242038/apps/anders-klanten-dev",
  "azp": "/projects/466589242038/apps/anders-klanten-dev",
  "email": "apigateway-3621-testing@appsembler-testing.iam.gserviceaccount.com",
  "exp": 1638291196,
  "iat": 1638290596,
  "iss": "https://cloud.google.com/iap",
  "sub": "accounts.google.com:109967078724472699637"
}
```

We still have an `email`, but there's no `hd`.

So we need to not crash when that happens. Luckily, we typically don't
want to grant the API Gateway service account elevated permissions, so
it's fine to just let the domain default to `None`.